pyproject.toml
[project]
name = "mlops-image-classification"
version = "0.1.0"
description = "MLOps project: Image Classification"
authors = [
    {name = "TM-Squared",email = "malaury.toussi@gmail.com"}
]
license = {text = "LICENCE"}
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "boto3 (>=1.37.28,<2.0.0)",
    "mysql-connector-python (>=9.2.0,<10.0.0)",
    "requests (>=2.32.3,<3.0.0)",
    "tensorflow[and-cuda] (>=2.19.0,<3.0.0)",
    "keras-squeezenet (>=0.4,<0.5)",
    "pillow (>=11.2.1,<12.0.0)",
    "mlflow (>=2.22.0,<3.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false

[tool.poetry.group.dev.dependencies]
python-dotenv = "^1.1.0"

